Ubuntu: Install `graphviz`
The Graphviz layout programs take descriptions of graphs in a simple
text language, and make diagrams in useful formats, such as images and
SVG for web pages; PDF or Postscript for inclusion in other documents;
or display in an interactive graph browser.

Nice for storing graphs in version control.

See http://www.graphviz.org/