bootstrap: Support `Brewfile`
The idea is to have a simple `Brewfile` for all OSX apps and use ansible
for the nifty details and for the Ubuntu setup.

The `Brewfile` is easier to manage than the `group_vars` approach and
keeps every Mac application in one place. With the `group_vars` I had to
store the package names in different locations.